Move from const to service.py
